regex: align Envoy API and xDS API
