+++ api/envoy/type/matcher/v3/regex.proto
@@ -2,6 +2,7 @@ syntax = "proto3";
 
 package envoy.type.matcher.v3;
 
+import "google/protobuf/any.proto";
 import "google/protobuf/wrappers.proto";
 
 import "envoy/annotations/deprecation.proto";
@@ -48,11 +49,22 @@ message RegexMatcher {
         [deprecated = true, (envoy.annotations.deprecated_at_minor_version) = "3.0"];
   }
 
+  message RegexEngine {
+    // The name of the engine.
+    string name = 1 [(validate.rules).string = {min_len: 1}];
+
+    // The typed config for the engine.
+    google.protobuf.Any typed_config = 2 [(validate.rules).any = {required: true}];
+  }
+
   oneof engine_type {
     option (validate.required) = true;
 
     // Google's RE2 regex engine.
-    GoogleRE2 google_re2 = 1 [(validate.rules).message = {required: true}];
+    GoogleRE2 google_re2 = 1;
+
+    // Typed regex engine.
+    RegexEngine engine = 3;
   }
 
   // The regex match string. The string must be supported by the configured engine.
