+++ test/common/common/matchers_test.cc
@@ -375,7 +375,7 @@ TEST(StringMatcher, SafeRegexValueIgnoreCase) {
                             "ignore_case has no effect for safe_regex.");
 }
 
-TEST(StringMatcher, XDSSafeRegexValue) {
+TEST(StringMatcher, SafeRegexValueFromTypedConfig) {
   const std::string yaml_string = R"EOF(
 safe_regex:
   engine:
@@ -384,7 +384,7 @@ safe_regex:
       "@type": type.googleapis.com/envoy.type.matcher.v3.RegexMatcher.GoogleRE2
   regex: foo.*
 )EOF";
-  xds::type::matcher::v3::StringMatcher matcher;
+  envoy::type::matcher::v3::StringMatcher matcher;
   TestUtility::loadFromYaml(yaml_string, matcher);
   EXPECT_TRUE(Matchers::StringMatcherImpl(matcher, ProtobufMessage::getStrictValidationVisitor())
                   .match("foo"));
@@ -392,24 +392,6 @@ safe_regex:
                   .match("foobar"));
   EXPECT_FALSE(Matchers::StringMatcherImpl(matcher, ProtobufMessage::getStrictValidationVisitor())
                    .match("bar"));
-}
-
-TEST(StringMatcher, XDSSafeRegexValueIgnoreCase) {
-  const std::string yaml_string = R"EOF(
-safe_regex:
-  engine:
-    name: google-re2
-    typed_config:
-      "@type": type.googleapis.com/envoy.type.matcher.v3.RegexMatcher.GoogleRE2
-  regex: foo.*
-ignore_case: true
-)EOF";
-  xds::type::matcher::v3::StringMatcher matcher;
-  TestUtility::loadFromYaml(yaml_string, matcher);
-  EXPECT_THROW_WITH_MESSAGE(
-      Matchers::StringMatcherImpl(matcher, ProtobufMessage::getStrictValidationVisitor())
-          .match("foo"),
-      EnvoyException, "ignore_case has no effect for safe_regex.");
 }
 
 TEST(PathMatcher, MatchExactPath) {
