+++ test/common/matcher/value_input_matcher_test.cc
@@ -17,16 +17,5 @@ TEST(ValueInputMatcher, TestMatch) {
   EXPECT_FALSE(matcher.match(absl::nullopt));
 }
 
-TEST(ValueInputMatcher, TestXDSMatch) {
-  xds::type::matcher::v3::StringMatcher matcher_proto;
-  matcher_proto.set_exact("exact");
-
-  StringInputMatcher matcher(matcher_proto, ProtobufMessage::getStrictValidationVisitor());
-
-  EXPECT_TRUE(matcher.match("exact"));
-  EXPECT_FALSE(matcher.match("not"));
-  EXPECT_FALSE(matcher.match(absl::nullopt));
-}
-
 } // namespace Matcher
 } // namespace Envoy
